Complete

+++ src/main/java/controller/DeleteCustomer.java
@@ -16,17 +16,16 @@ public class DeleteCustomer extends HttpServlet {
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		if (req.getSession().getAttribute("admin") == null) {
-			resp.getWriter().print("<h1 style='color:red'>Invalid Session</h1>");
+			resp.getWriter().print("<html><h2>Invalid Session</h2>");
 			req.getRequestDispatcher("Login.html").include(req, resp);
 		} else {
 
-			// Getting Id from the URL
 			int id = Integer.parseInt(req.getParameter("id"));
 			MyDao dao = new MyDao();
-			// Finding object because remove method accepts object
 			Customer customer = dao.findCustomer(id);
+			if(customer!=null)
 			dao.delete(customer);
-			resp.getWriter().print("<h1 style='color:green'>Data Deleted Successfully</h1>");
+			resp.getWriter().print("<html><h3>Data Deleted Successfully</h3>");
 			req.getRequestDispatcher("viewcustomer").include(req, resp);
 		}
 	}
